Separate and parallelize Go repository tests in CircleCI.
Go repo tests are now placed into a separate subpackage and executed in
a parallelized job separately from regular go tests. The test now
detects when it runs under CircleCI and shards itself accordingly.
However, using `go test ./...` still works as before, which is handy
for local development.
