+++ tests/vendored_test.go
@@ -7,28 +7,6 @@ import (
 	"testing"
 )
 
-// Go repository basic compiler tests, and regression tests for fixed compiler bugs.
-func TestGoRepositoryCompilerTests(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping Go repository tests in the short mode")
-	}
-	if runtime.GOARCH == "js" {
-		t.Skip("test meant to be run using normal Go compiler (needs os/exec)")
-	}
-
-	args := []string{"go", "run", "run.go", "-summary"}
-	if testing.Verbose() {
-		args = append(args, "-v")
-	}
-	cmd := exec.Command(args[0], args[1:]...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stdout
-	err := cmd.Run()
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 // Test that GopherJS can be vendored into a project, and then used to build Go programs.
 // See issue https://github.com/gopherjs/gopherjs/issues/415.
 func TestGopherJSCanBeVendored(t *testing.T) {
